+++ packages/model/src/index.ts
@@ -1,1 +1,4 @@
+export * from "./client.js";
 export * from "./reasoning.js";
+export * from "./sse.js";
+export * from "./stream-events.js";